refactor: :art: group files by feature
