+++ src/features/data_processing/ObjectiveInsert.js
@@ -1,8 +1,8 @@
 import { useContext, useEffect, useState } from "react";
-import {AppContext} from "./AppContext";
-import { getExams, getExamSubjects, getUnitModules, getSubjectUnits, getModuleTopics, addObjective } from "./services/SQLService";
+import {AppContext} from "../../AppContext";
+import { getExams, getExamSubjects, getUnitModules, getSubjectUnits, getModuleTopics, addObjective } from "../../services/SQLService";
 import './CSS/ObjectiveInsert.css';
-import DependentDropdown from "./DependentDropdown";
+import DependentDropdown from "../../DependentDropdown";
 
 const ObjectiveInsert = () => {
