+++ src/features/pastpapers/Questions.js
@@ -1,9 +1,9 @@
 import { useEffect, useState, useContext } from "react";
 import { Link, useParams } from "react-router-dom"
-import { getPaperQuestions, updateUserQuestion } from "./services/SQLService";
-import FamiliarityDropdown from "./FamiliarityDropdown";
-import CompleteCheckbox from "./CompleteCheckbox";
-import {AppContext} from "./AppContext";
+import { getPaperQuestions, updateUserQuestion } from "../../services/SQLService";
+import FamiliarityDropdown from "../../FamiliarityDropdown";
+import CompleteCheckbox from "../../CompleteCheckbox";
+import {AppContext} from "../../AppContext";
 import './CSS/Questions.css';
 import './CSS/global.css'
