+++ src/features/quiz/QuizPreview.js
@@ -1,4 +1,4 @@
-import { useQuizContext } from "./contexts/QuizContext";
+import { useQuizContext } from "../../contexts/QuizContext";
 
 const QuizPreview = () => {
     // get topics
